// miniprogram/pages/index/cpns/TodayTomorrow/type.ts
export type weatherType = {
  aqi_level: number
  aqi_name: string
  aqi_url: string
  day_weather: string
  day_weather_code: string
  day_weather_short: string
  day_weather_url: string
  day_wind_direction: string
  day_wind_direction_code: string
  day_wind_power: string
  day_wind_power_code: string
  max_degree: string
  min_degree: string
  night_weather: string
  night_weather_code: string
  night_weather_short: string
  night_weather_url: string
  night_wind_direction: string
  night_wind_direction_code: string
  night_wind_power: string
  night_wind_power_code: string
  time: string
} 